// types.ts
export interface Expense {
    id: string;
    parentId?: string;
    date: string;
    category: string;
    payee: string;
    amount: number;
    status: 'Paid' | 'Unpaid';
}

export interface Settlement {
    id: string;
    date: string;
    amount: number;
    type: 'Bank Transfer' | 'Cash Deposit' | 'Card Settlement' | 'Internal Transfer';
    fromAccount: string;
    toAccount: string;
}

export interface Account {
  id: string;
  parentId?: string;
  name: string;
  type: 'Asset' | 'Liability' | 'Equity' | 'Revenue' | 'Expense';
  subType: string;
  balance: number;
  description: string;
}